app/processing_service/requirements.txt
# API и веб-сервер
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Базы данных
clickhouse-driver==0.2.6
clickhouse-connect==0.7.0
psycopg2-binary==2.9.9

# Kafka
confluent-kafka==2.3.0

# Prometheus
prometheus-client==0.19.0
aiohttp==3.9.1
aiokafka==0.10.0
prometheus-async==19.2.0